requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
cffi==1.17.1
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.12
decorator==5.1.1
Django==5.1.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
executing==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
PyWavelets==1.8.0
pyzmq==26.2.0
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
# pyzbar opencv-python pyzxing sql
# pip install sqlalchemy
# sudo apt-get install libzbar0